+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Prowide ISO 20022 - CHANGELOG
 
 #### 9.5.6 - SNAPSHOT
-  * Changed the `MxParseUtils` findByTags and findByPath methods to return the element values and not the XML stream object
+  * Changed the `MxParseUtils` findByTags and findByPath methods to return the element values instead of the XML stream object
 
 #### 9.5.5 - January 2025
   * (PW-2113) `MxParseUtils` added a method to extract the enclosed MT from a multi-format MX message